feat: onHitSpecificBullet 함수 인자 추가

+++ Source/DefenceGame/Private/PooledBullet.cpp
@@ -19,7 +19,7 @@ void APooledBullet::Tick(float DeltaTime)
 	SetActorLocation(GetActorLocation() + GetActorForwardVector() * speed * DeltaTime);
 }
 
-float APooledBullet::GetDamage() const
+int32 APooledBullet::GetDamage() const
 {
 	return damage;
 }
